backend/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
anthropic
langchain
langchain-community
faiss-cpu
sentence-transformers
transformers
PyPDF2
pypdf
torch
pydantic